requirements.txt
GitPython
PyGitHub
graphviz